support for read EPROMS 2764-27128-27256-27512, with linux client

+++ Arduino/AT28C_Programmer/SRHelper.cpp
@@ -45,7 +45,7 @@ void shiftOutFast(byte data) {
   }
 }
 
-void addressWrite(unsigned int value) {
+void addressWrite(unsigned long value) {
   byte b1 = 0;
   byte b2 = 0;
